Fix the four build warnings, so that a new one is visible
Two of them are the price of LLVM's C interface, which takes a debug-info
column and an extract-value index as plain unsigned while conec produces
a ptrdiff_t and a size_t. The column gets an explicit cast at the
boundary; the index is fixed at its declaration instead, since feeding
LLVMBuildExtractValue is the only thing it does.

The third is Windows-specific: %lu is unsigned long, which stays 32-bit
on MSVC x64 while size_t is 64, so the memory figure in the compile
summary was formatted with a specifier that cannot hold it. The same line
is fine on Linux, where long is 64-bit. It is the only one of the four
that was undefined rather than merely lossy.

The fourth is an int local holding a copy of a size_t table size, which
then supplies a memcpy length.

None of them could fire. Reaching one needs a source line of four billion
characters, a tuple of four billion elements, a flow table past two
billion entries, or four terabytes of tracked memory. The reason to fix
them is that four is the worst count to sit at: too few to be worth a
project, and too many for a new warning to stand out against. The test
suite already holds every Cone source in the corpus to zero warnings;
this holds the compiler's own build to the same bar.

Verified by a from-scratch 82-unit build, which now reports none, and by
a --debug compile, which is the only path that reaches the column cast.
No design note changes: nothing here alters compiler behavior, and no
scenario shape exists for asserting a clean C build.

+++ src/c-compiler/genllvm/genlexpr.c
@@ -956,7 +956,7 @@ void genlStore(GenState *gen, INode *lval, LLVMValueRef rval) {
 LLVMValueRef genlExpr(GenState *gen, INode *termnode) {
     if (!gen->opt->release && gen->fn) {
         LLVMMetadataRef loc = LLVMDIBuilderCreateDebugLocation(gen->context, 
-            termnode->linenbr, termnode->srcp-termnode->linep, LLVMGetSubprogram(gen->fn), NULL);
+            termnode->linenbr, (unsigned)(termnode->srcp-termnode->linep), LLVMGetSubprogram(gen->fn), NULL);
         LLVMValueRef val = LLVMMetadataAsValue(gen->context, loc);
         LLVMSetCurrentDebugLocation(gen->builder, val);
     }
@@ -1089,7 +1089,7 @@ LLVMValueRef genlExpr(GenState *gen, INode *termnode) {
             TupleNode *tuple = (TupleNode*)reftype;
             INode **nodesp;
             uint32_t cnt;
-            size_t index = 0;
+            unsigned index = 0;
             int16_t *countp = anode->counts;
             for (nodesFor(tuple->elems, cnt, nodesp)) {
                 if (*countp != 0) {
